+++ samples/NotifyGen.AvaloniaSample/Address.cs
@@ -1,0 +1,8 @@
+namespace NotifyGen.AvaloniaSample;
+
+[Notify]
+public partial class Address
+{
+    private string _city = "Berlin";
+    private string _country = "Germany";
+}